Move out the logic getting rewrite/recipe classpaths
